add pg_wrapper to control and monitor PG

+++ src/k2/connector/yb/common/stringprintf.cc
@@ -131,10 +131,10 @@ string StringPrintfVector(const char* format, const vector<string>& v) {
   // or displaying random chunks of memory to users.
 
   const char* cstr[kStringPrintfVectorMaxArgs];
-  for (long unsigned int i = 0; i < v.size(); ++i) {
+  for (int i = 0; i < v.size(); ++i) {
     cstr[i] = v[i].c_str();
   }
-  for (long unsigned int i = v.size(); i < arraysize(cstr); ++i) {
+  for (int i = v.size(); i < arraysize(cstr); ++i) {
     cstr[i] = &string_printf_empty_block[0];
   }
